Add `salloc` and `sbatch`, rework `mila code`

Add `sbatch` for RemoteV2


Add better docstrings


Move `code` command to code_command, add new impl


Fix circular import issue


Fix import of `code_v1`


Use new `code` impl in integration test


Fix bug in `is_already_logged_in`


Adjust `mila code` integration test


Add mark to tests that launch jobs


Fix typing error for py38


Fix typing error for py38


Fix missing imports in common.py
