+++ pyproject.toml
@@ -40,6 +40,7 @@ pytest-mock = "^3.11.1"
 pytest-socket = "^0.6.0"
 pytest-cov = "^4.1.0"
 pytest-timeout = "^2.2.0"
+pytest-asyncio = "^0.23.6"
 
 
 [tool.isort]
