+++ tests/utils/test_remote_v2.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pathlib import Path
 from unittest.mock import Mock
 
@@ -5,6 +7,8 @@ import pytest
 
 import milatools.utils.remote_v2
 from milatools.utils.remote_v2 import (
+    InteractiveRemote,
+    PersistentRemote,
     RemoteV2,
     UnsupportedPlatformError,
     control_socket_is_running,
@@ -13,7 +17,10 @@ from milatools.utils.remote_v2 import (
 )
 from tests.integration.conftest import skip_param_if_not_already_logged_in
 
-from ..cli.common import requires_ssh_to_localhost, xfails_on_windows
+from ..cli.common import (
+    requires_ssh_to_localhost,
+    xfails_on_windows,
+)
 
 pytestmark = [xfails_on_windows(raises=UnsupportedPlatformError, strict=True)]
 
@@ -85,3 +92,53 @@ def test_is_already_logged_in(
 def test_controlsocket_is_running(cluster: str, already_logged_in: bool):
     control_path = get_controlpath_for(cluster)
     assert control_socket_is_running(cluster, control_path) == already_logged_in
+
+
+# make it last a bit longer here so we don't confuse end of command/test with end of job.
+@pytest.mark.parametrize("allocation_flags", [{"time": "00:01:00"}], indirect=True)
+def test_salloc(login_node_v2: RemoteV2, allocation_flags: list[str]):
+    compute_node = login_node_v2.salloc(allocation_flags)
+    assert isinstance(compute_node, InteractiveRemote)
+    assert compute_node.hostname != login_node_v2.hostname
+
+    job_id = compute_node.get_output("echo $SLURM_JOB_ID")
+    assert job_id.isdigit()
+    assert compute_node.job_id == int(job_id)
+
+    slurm_env_vars = {
+        (split := line.split("="))[0]: split[1]
+        for line in compute_node.get_output("env | grep SLURM").splitlines()
+    }
+    # NOTE: We don't yet have all the other SLURM env variables here yet because we're
+    # only ssh-ing into the compute node.
+    assert slurm_env_vars == {"SLURM_JOB_ID": str(compute_node.job_id)}
+
+    all_slurm_env_vars = {
+        (split := line.split("="))[0]: split[1]
+        for line in compute_node.get_output("srun env | grep SLURM").splitlines()
+    }
+    assert all_slurm_env_vars["SLURM_JOB_ID"] == str(compute_node.job_id)
+    assert len(all_slurm_env_vars) > 1
+
+
+def test_sbatch(login_node_v2: RemoteV2, allocation_flags: list[str]):
+    compute_node = login_node_v2.sbatch(allocation_flags)
+    assert isinstance(compute_node, PersistentRemote)
+
+    assert compute_node.hostname != login_node_v2.hostname
+    job_id = compute_node.get_output("echo $SLURM_JOB_ID")
+    assert job_id.isdigit()
+    assert compute_node.job_id == int(job_id)
+    # Same here, only get SLURM_JOB_ID atm because we're ssh-ing into the node.
+    slurm_env_vars = {
+        (split := line.split("="))[0]: split[1]
+        for line in compute_node.get_output("env | grep SLURM").splitlines()
+    }
+    assert slurm_env_vars == {"SLURM_JOB_ID": str(compute_node.job_id)}
+
+    all_slurm_env_vars = {
+        (split := line.split("="))[0]: split[1]
+        for line in compute_node.get_output("srun env | grep SLURM").splitlines()
+    }
+    assert all_slurm_env_vars["SLURM_JOB_ID"] == str(compute_node.job_id)
+    assert len(all_slurm_env_vars) > 1
